move legacy payment processing to its own class
