+++ src/Payments/Api.php
@@ -7,9 +7,6 @@ use Automattic\WooCommerce\Blocks\Payments\Integrations\BankTransfer;
 use Automattic\WooCommerce\Blocks\Payments\Integrations\CashOnDelivery;
 use Automattic\WooCommerce\Blocks\Payments\Integrations\Cheque;
 use Automattic\WooCommerce\Blocks\Payments\Integrations\PayPal;
-use Automattic\WooCommerce\StoreApi\Payments\PaymentContext;
-use Automattic\WooCommerce\StoreApi\Payments\PaymentResult;
-use Automattic\WooCommerce\StoreApi\Utilities\NoticeHandler;
 
 /**
  *  The Api class provides an interface to payment method registration.
@@ -52,7 +49,6 @@ class Api {
 		add_action( 'woocommerce_blocks_checkout_enqueue_data', array( $this, 'add_payment_method_script_data' ) );
 		add_action( 'woocommerce_blocks_cart_enqueue_data', array( $this, 'add_payment_method_script_data' ) );
 		add_action( 'woocommerce_blocks_payment_method_type_registration', array( $this, 'register_payment_method_integrations' ) );
-		add_action( 'woocommerce_rest_checkout_process_payment_with_context', array( $this, 'process_legacy_payment' ), 999, 2 );
 		add_action( 'wp_print_scripts', array( $this, 'verify_payment_methods_dependencies' ), 1 );
 	}
 
@@ -118,57 +114,6 @@ class Api {
 		$payment_method_registry->register(
 			Package::container()->get( CashOnDelivery::class )
 		);
-	}
-
-	/**
-	 * Attempt to process a payment for the checkout API if no payment methods support the
-	 * woocommerce_rest_checkout_process_payment_with_context action.
-	 *
-	 * @param PaymentContext $context Holds context for the payment.
-	 * @param PaymentResult  $result  Result of the payment.
-	 */
-	public function process_legacy_payment( PaymentContext $context, PaymentResult &$result ) {
-		if ( $result->status ) {
-			return;
-		}
-
-		// phpcs:ignore WordPress.Security.NonceVerification
-		$post_data = $_POST;
-
-		// Set constants.
-		wc_maybe_define_constant( 'WOOCOMMERCE_CHECKOUT', true );
-
-		// Add the payment data from the API to the POST global.
-		$_POST = $context->payment_data;
-
-		// Call the process payment method of the chosen gateway.
-		$payment_method_object = $context->get_payment_method_instance();
-
-		if ( ! $payment_method_object instanceof \WC_Payment_Gateway ) {
-			return;
-		}
-
-		$payment_method_object->validate_fields();
-
-		// If errors were thrown, we need to abort.
-		NoticeHandler::convert_notices_to_exceptions( 'woocommerce_rest_payment_error' );
-
-		// Process Payment.
-		$gateway_result = $payment_method_object->process_payment( $context->order->get_id() );
-
-		// Restore $_POST data.
-		$_POST = $post_data;
-
-		// If `process_payment` added notices, clear them. Notices are not displayed from the API -- payment should fail,
-		// and a generic notice will be shown instead if payment failed.
-		wc_clear_notices();
-
-		// Handle result.
-		$result->set_status( isset( $gateway_result['result'] ) && 'success' === $gateway_result['result'] ? 'success' : 'failure' );
-
-		// set payment_details from result.
-		$result->set_payment_details( array_merge( $result->payment_details, $gateway_result ) );
-		$result->set_redirect_url( $gateway_result['redirect'] );
 	}
 
 	/**
